backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pandas==2.2.3
openpyxl==3.1.5
openai==1.54.3
faiss-cpu==1.9.0
numpy==1.26.4
pydantic==2.9.2
httpx==0.27.2
yfinance==0.2.48
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
sentence-transformers==3.3.1
